move algorithms to new package
